Make some of the TraceView code re-usable.
Factoring it out into a base class.
